Update Issue 264 Label:-VerifiedFilter
We use the Chrome Element Inspector (in the right click menu) and figure
out what ids and classes are used to identify the ad, then use CSS selector
rules to specify them.

It works in FF either because of an unknown bug in AdBlock (that I'm trying
to find) or because Chrome doesn't support blocking of javascript yet, and
the FF rule blocks a javascript.


svn path=/trunk/; revision=91