Followup on r292:
Since the page isn't reloaded when a filter is removed, we don't need
to store the edited filter in a cookie -- we can store it temporarily
in a global variable.  This also gets rid of the need to escape() and
unescape(), and to parse out the string parts of the cookie.


svn path=/trunk/; revision=297