Update Issue #3579
True resource blocking in Chrome 6.  Chrome doesn't fully support this, so a
few resources might still get downloaded before my code has a chance to get
itself in the way of the download.  Therefore I still do an ad hiding sweep on
the page as usual to catch anybody who slipped through the resource blocks.

I had to do this uglily -- we load the entirety of our filters in every content
script -- which adds about 100ms to page load time.  This should be more than
made up for by the savings in ad downloads.  Users can turn off the option
(it's on by default) in Options -> General.

Stable Chrome users see a big red note saying they have to upgrade before the
feature will work for them.  Safari users don't see the option in the Options
page, because it's always turned on for them, and done properly.


svn path=/trunk/; revision=579