Update issue 271 Status: Started Labels: ReadyForRelease
Add a collapse blocked elements function. This is a completely different one than the previous, as the previous did match CSS selectors and this one removes the empty space if an element resource was blocked.

If you merge it: Do only merge this one and not the previous!!!

Please also check this on Safari before releasing!

svn path=/branches/issue-271_2/; revision=666