Update issue 3261 Owner: fam.lam@live.nl Status: Started Labels: ReadyForRelease
AdBlock now follows the expires: or expires after from a list. It does ignore lists that would like to be updated after a number of hours that are not a multiple of 24 (they are updated when the number of hours reaches a multiple of 24.

It removes the 'auto-update on startup after 5 days, as that would make this feature useless.
It also doesn't check the full list, it only checks the first 15 lines for an 'expires' line. That should be enough in my opinion...

svn path=/branches/issue-3261/; revision=683