Update issue 3751 Labels: ReadyForRelease
Funny... I don't see anything flash using your repro steps. However, you were right that removing that part of the code still subscribes it to EasyList. Nice side-effect. That is because I first implemented it into the options page and after that in the abp:// link subscription + the 'subscribe to a custom filter list' textbox. 
However, according to your comment it should be fixed now, so I am re-adding the ReadyForRelease label.

svn path=/branches/issue-3751/; revision=697