Update issue 3879 There are two things going on here: 1) AdBlock adds a fixed amount of RAM to store filters on the page because Chrome doesn't support synchronous calls to the background page yet 2) AdBlock has to sweep the page for ads every time a node is inserted, because Chrome doesn't support synchronous calls to the background page yet so some ad scripts could slip through
1) is a problem on TechCrunch because TechCrunch loads dozens of subframes,
and each one loads filters in order to emulate synchronous calls to the
background.

This code change fixes #1 by only loading filters in the top level frame --
subframes do not do 'true blocking'.  This should be OK in almost all cases,
since not a lot of sites load a subframe that has real content and also
loads ads.

It doesn't address the slow leaking of RAM each time a node is inserted,
but I'm going to bump this down to pri-high again until we hear that
the slow leaking is actually a problem.  I think people just noticed web
pages with lots of subframes, causing lots of memory usage.

svn path=/branches/issue-3879/; revision=819