Update issue 4180 Correct a bug in Safari's true blocking: it was waiting for an async call to the background in order to see if the page was whitelisted, before turning on blocking.  Which means that resources that loaded quickly enough were not getting blocked!
Also, Safari was storing the filtertext needlessly on target pages (due to some code for
Chrome to emulate true blocking), which was wasting memory.  Hopefully this will fix the
reports of large memory usage in Safari.

Finally, I realized that Chrome doesn't have to store *all* filters on every target page;
it can just store the filters that apply to that domain.  It knocks out about 40% of them,
which should reduce the extra memory used by Chrome significantly.  And of course, once
Chrome supports true blocking, we can get rid of this hack and not use any extra memory
per target page.

svn path=/branches/issue-4180/; revision=846