Update issue 3768 Show a badge whenever there's a new version worth talking about. Don't show the badge to users who just installed, though. Show a 'what's new' div in the toolbar button when there's a new version worth talking about. The badge disappears upon first toolbar button click. The div disappears when the user dismisses it.
We should only display a badge for relatively large new features so we don't annoy the user with badges.

Also, rearrange and reword the popup menu items somewhat.

Finally, I'm updating the manifest to 2.2.0 now, since the 'new version' notification value is 2.2... to remind myself that we need to up the minor version number when this is released.  At that point, all old users who had to go through the Chrome Re-Enable will see the badge.

svn path=/branches/issue-3768-prime/; revision=1150