Update issue 5029 Status: Fixed Merged into trunk
Good job on handling ~option better than my original code did.
I don't know that we should keep ElementType.donottrack, though, until
we can do something with it (is there a reason to?)

Anyway, for the sake of expediency, and since you've tested this, I'll
go ahead and release as-is.  Thanks for the fix.


svn path=/trunk/; revision=1613