update issue 5574 Status:Accepted Cc:john.rey92 Okay, so based off Michael's feedback on the dev groups page thing, I think we can take care of Issue 5441 (since it has stalled) if we include a link to either the specific Issues for translations or the HowToTranslate wiki. If we choose the latter, my suggestion would be to mark all the translation issues under a common header so we can link to that as a search. For example: http://code.google.com/p/adblockforchrome/issues/list?q=label:complete-translation
Or something to that effect (as long as you understand my general idea). The user can just click through each link then which will make it easier hopefully for them, at least that is the idea of Issue 5574.

I reverted the changes I made before, and added a link to the German translation for the wiki page. I also added more to the explanation saying to leave the link and to not include any external links. I'd obviously have to manually link each translation (oh fun times).

Unless there is a better way of doing it I am open to that as well :)

svn path=/branches/issue-5574/; revision=2343