Update issue 5924
> on nirsoft.net with EasyList enabled, sometimes ad remains aren't removed correctly, just above "What's new"

The workaround.  Blocked elements are now removed even when onBeforeRequest
fires before beforeload.

There will be small memory leaks when multiple frames on the same tab have the
same URL.  The leak is on the order of

  100 bytes * N * (N-1) * M

where N is the number of frames in a tab that share a URL and
      M is the number of image+iframe+object resources loaded by that URL.

For example, if 3 frames load a URL that has 30 images, 10 objects, and 10
iframes, that's 100B * 3 * 2 * (30+10+10) = 30kB leaked.  The leak is in
the content scripts, so it will be freed on tab navigate/close.


svn path=/branches/issue-5924/; revision=2998