Update issue 6067 Status: NeedsReview For some reason I am now able to subscribe to a few more lists before I reach the limit. However, the bug still exists: once you are over the limit, you have to clear your settings the hard way, which isn't the way it should happen.
This revision fixes that: when you exeed the limit, you'll be send to the options page with the request to unsubscribe. There is no way to get around this message (as AdBlock for example wouldn't be able to update the lists anymore, you have to take action immediately).
I know there is a bug that when you are already on the options page, it still opens a second options page tab, but I don't care about that :).

Also removes some TEMP code which is old enough to be deleted.

svn path=/branches/issue-6067/; revision=3048