update issue 6009 Status: NeedsReview
Update the filters when the computer is idle rather than when it's in use.
There is also the possibility to specify a maximum timeout, after which the function triggers anyway, idle or not. (So a user who always closes his browser when he is just 10 seconds away still gets updates).

The idle permission is a permission without warnings, so adding it shouldn't hurt anyone.

idlehandler can run on the options page as well (maybe even in the content scripts, but I didn't test that). I've put it in a separate file, so that it can easily be used in other places if that's what we want. However, in those cases we should be aware that there is a forced minimum of 15 seconds before idle triggers...

I didn't add _updateSubscriptionText to the idle running functions, as I don't want to risk (worst case scenario) that both the _myfilters._subscriptions contain .text and for some reason all other subscriptions updated at the same time and these exist in the memory too, all waiting for idle, which can take up to 10 minutes. Who knows what function triggers in the meantime. How to handle unsubscribing+resubscribing etcetera? So I decided not to implement that.

svn path=/branches/issue-6009/; revision=3075