Update issue 5937 Status: NeedsReview Address c21 and c22
The problem: in 'new' style, resourceblock was using the background's
_myfilters instead of building its own local_filterset.  But the background
caches boolean results, not filter text results, in its filterset._matchCache,
so when resourceblock checked whether a resource was whitelisted, it got back a
'false' instead of a whitelist filtertext.

After webRequest is in stable, I'll change the matchCache to store Filter objects
instead of booleans, and resourceblock can use _myfilters directly.

svn path=/branches/issue-5937/; revision=3190