Update issue 6360 Status: NeedsReview
Add an onReady function.  Famlam, please set to RFR when ready.
The thing is, I'm not sure after all that this is needed, because I
can't get debug_print_selector_matches() to fail WITHOUT my change.
But in any case, I believe my code can't harm anything, and has the
possibility of making things run a little sooner, so I think it's
OK to put in.

Verified that debug_print_selector_matches and abp:subscribe links
work properly in Chrome 17 and Chrome 18.

svn path=/branches/issue-6360/; revision=3442