Update issue 6430 Status: NeedsReview [Note: subbranch of 6758 because that branch fixes a related issue, which would make testing more difficult]
Send a new event for custom list updates.
I did not send the 'filters_updated' event, because that would also trigger on options/filters and subscribe, while it shouldn't. Neither should options/customize trigger on list updates.
I verified the text box on customize gets updated when 
1. resourceblock adds/removes a filter
2. the blacklister adds a filter
3. the button adds or removes a whitelist
And that it doesn't get updated when it is editable (that is, when the user clicked the 'edit' button).

svn path=/branches/issue-6430/; revision=3876