Update issue 5626 Thanks! I've uploaded the (second) translation. It'll be in the next version.
Are you starred to the 'HowToTranslate' page? If so, I think I'll start adding a comment each time we add a new string. (Of course, there may be a delay for a few days as I often forget to update that page, but well, a few days shouldn't matter that much). That way you (and other translators) can get notified about translation updates.

svn path=/trunk/; revision=4239