Update  Issue 4899 Status: NeedsReview cc: adblockforchrome
Added a flag to control the binding. This will prevent binding the enter keypress to be bonded more than once. Also changed default_button to adblock_default_button to avoid colliding with page buttons.

To test choose don't run on pages on this domain and use enter instead of buttons, then instead of clicking the buttons, use enter instead.

svn path=/branches/issue-4899/; revision=5088