Update Issue 4538 Thanks!
Well, the best you can do at the moment is star this page: https://code.google.com/p/adblockforchrome/wiki/HowToTranslate
When there are new strings, someone will usually leave a comment, and people who starred the page receive an email.

That being said, updates are usually released very quickly, so it can be a bit difficult to get the translations ready on time. Hopefully in the future it'll be more straightforward, as described in issue 7427.

svn path=/trunk/; revision=5110