Update  Issue 5177
Fixed entering url to the upload box when an existing filter list already exist to work as expected. Removed Fetching label on page init if the list is already subscribed on. Change variables that ends with filter to end filter_list, also for Class Names and methods to end with FilterList. Remove prompt when there is a page update.

svn path=/branches/issue-5177/; revision=5231