Update  Issue 5950 Status: NeedsReview cc: adblockforchrome
Added checking of url to see if the user submitted filter is now a official subscription and vice versa.
Did not include values for apindo since it is only used for testing.
To test, please report by fam.lam@live.nl in this issue.

Requireslist issue(and QUnit if needed) to follow.

svn path=/branches/issue-5950/; revision=5327